data-processing/requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
transformers==4.36.2
torch==2.1.2

# AI and ML
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20

# Vector databases
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18
faiss-cpu==1.7.4

# Data processing utilities
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.2

# Text processing
regex==2023.12.25
tiktoken==0.5.2
sentence-transformers==2.2.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Database connections
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# API and web
fastapi==0.108.0
uvicorn==0.25.0
pydantic-settings==2.1.0

# Utilities
tqdm==4.66.1
python-multipart==0.0.6
httpx==0.26.0
